+++ src/components/ui/MsgPopover.tsx
@@ -2,8 +2,10 @@
 
 import React from 'react';
 
-import Popover from '@mui/material/Popover';
+import Popper from '@mui/material/Popper';
 import Typography from '@mui/material/Typography';
+import Paper from '@mui/material/Paper';
+import Fade from '@mui/material/Fade';
 
 import CheckBoxOutlinedIcon from '@mui/icons-material/CheckBoxOutlined';
 
@@ -13,8 +15,16 @@ const MsgPopover = (props: {
 	anchorEl: HTMLElement | null;
 	onClose: () => void;
 	closeAfter?: number;
+	transitionDuration?: number | { enter?: number; exit?: number };
 }) => {
-	const { msg, open, anchorEl, onClose, closeAfter = 600 } = props;
+	const {
+		msg,
+		open,
+		anchorEl,
+		onClose,
+		closeAfter = 1000,
+		transitionDuration = { enter: 300, exit: 150 },
+	} = props;
 
 	React.useEffect(() => {
 		if (open) {
@@ -24,27 +34,32 @@ const MsgPopover = (props: {
 	}, [open, onClose, closeAfter]);
 
 	return (
-		<Popover
+		<Popper
 			open={open && Boolean(anchorEl)}
 			anchorEl={anchorEl}
-			anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-			transformOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-			transitionDuration={100}
+			placement="top"
 			aria-hidden={!open}
+			transition
 		>
-			<div
-				style={{
-					display: 'flex',
-					flexDirection: 'row',
-					alignItems: 'center',
-					gap: 4,
-					padding: 2,
-				}}
-			>
-				<Typography sx={{ p: 1 }}>{msg}</Typography>
-				<CheckBoxOutlinedIcon color="success" />
-			</div>
-		</Popover>
+			{({ TransitionProps }) => (
+				<Fade {...TransitionProps} timeout={transitionDuration || 350}>
+					<Paper
+						elevation={3}
+						sx={{
+							display: 'flex',
+							flexDirection: 'row',
+							alignItems: 'center',
+							gap: 2,
+							padding: 1,
+							borderRadius: 2,
+						}}
+					>
+						<Typography sx={{ p: 1 }}>{msg}</Typography>
+						<CheckBoxOutlinedIcon color="success" />
+					</Paper>
+				</Fade>
+			)}
+		</Popper>
 	);
 };